+++ __fxstat.c
@@ -18,5 +18,6 @@ int __fxstat(int ver, int fd, struct stat *statbuf)
 {
 	__debug("%s(fd: %i <-> '%s', ...)\n", __func__, fd, __fpath(fd));
 
-	return __fxstatat(ver, fd, "", statbuf, AT_EMPTY_PATH);
+	return __fxstatat(ver, fd, "", statbuf,
+			  AT_EMPTY_PATH | AT_SYMLINK_NOFOLLOW);
 }